move election tests into a seperate file
